refactor: move bottomSection and topSection to main folder (Wen, Nov 1, 2023 23:39)
