+++ react-app/src/components/main/bottomSection/bottomSection.scss
@@ -7,7 +7,7 @@
   flex: 1 0 auto;
 
   .card {
-    flex-basis: 18%;
+    flex-basis: 18%
   }
 }
