+++ react-app/src/components/main/topSection/topSection.scss
